requirements.txt
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask==3.1.2
Werkzeug==3.1.3
markdown==3.9
python-slugify==8.0.4